fix Label relative paths after renaming the file
